+++ app/login.tsx
@@ -5,33 +5,19 @@ import {
   TextInput,
   TouchableOpacity,
   Platform,
-  SafeAreaView,
   ScrollView,
-  ScrollViewBase,
   KeyboardAvoidingView,
   StyleSheet,
-  Button,
   Image,
 } from "react-native";
 import { useForm, Controller } from "react-hook-form";
 import tw from "../lib/tailwind";
 import Logo from "../assets/images/logo.svg";
 import { Svg, Path, G, Mask } from "react-native-svg";
-import Google from "../assets/images/google-icon.png";
-import Fb from "../assets/images/fb-icon.png";
 import { Link, useRouter } from "expo-router";
-import {
-  signInWithPopup,
-  GoogleAuthProvider,
-  signInWithRedirect,
-  signInWithCredential,
-} from "firebase/auth";
-import { auth } from "../utils/firebase";
-// import * as Google from 'expo-google-app-auth';
-// import firebase from '../firebase';
-import firebase from "firebase/app";
-import "firebase/auth";
-import * as AuthSession from "expo-auth-session";
+import GoogleButton from "../components/Home/GoogleButton";
+import FacebookButton from "../components/Home/FacebookButton";
+
 
 type FormData = {
   firstName: string;
@@ -57,67 +43,6 @@ const Login = () => {
     router.push("/home");
   };
 
-  // const handleGoogleSignIn = async () => {
-  //   try {
-  //     const provider = new GoogleAuthProvider();
-  //     // await signInWithPopup(auth, provider);
-  //     const result = await signInWithPopup(auth, provider);
-
-  //   // Access the user information
-  //   const user = result.user;
-  //   const { displayName, email, photoURL } = user;
-  //   console.log(user)
-  //     // Sign-in successful, navigate to the desired page
-  //     // console.log(auth, provider)
-  //     router.push("/home");
-  //   } catch (error) {
-  //     // Handle sign-in error
-  //     console.error("Google sign-in error:", error);
-  //   }
-  // };
-
-
-
-  const handleGoogleSignIn = async () => {
-    try {
-      const provider = new GoogleAuthProvider();
-  
-      // Set additional scopes if needed
-      if (Platform.OS === "android") {
-        provider.addScope("profile");
-        provider.addScope("email");
-      }
-  
-      // Handle the sign-in flow based on the platform
-      if (Platform.OS === "web") {
-        const result = await signInWithPopup(auth, provider);
-        // Access the user information
-        const user = result.user;
-        const { displayName, email, photoURL } = user;
-        console.log("User info:", { displayName, email, photoURL });
-      } else {
-        const redirectUrl = AuthSession.makeRedirectUri({ useProxy: true });
-        const providerId = provider.providerId;
-        const authUrl = `https://${auth.config.authDomain}/__/auth/handler?provider=${providerId}&redirect_uri=${encodeURIComponent(
-          redirectUrl
-        )}`;
-        const result = await AuthSession.startAsync({ authUrl });
-        if (result.type === "success") {
-          // Complete the sign-in flow using the obtained accessToken
-          const { accessToken } = result.params;
-          await signInWithCredential(
-            auth,
-            GoogleAuthProvider.credential(null, accessToken)
-          );
-          // Sign-in successful, navigate to the desired page
-        }
-      }
-      router.push("/home");
-    } catch (error) {
-      // Handle sign-in error
-      console.error("Google sign-in error:", error);
-    }
-  };
 
   return (
     <ScrollView style={[tw`bg-white`, { fontFamily: "OpenSans_400Regular" }]}>
@@ -312,27 +237,11 @@ const Login = () => {
           </View>
 
           <View style={tw`flex-row w-full gap-x-4`}>
-            <TouchableOpacity
-              style={[
-                tw`py-4 px-6 shadow-md rounded-xl flex-1 w-full flex-row bg-white items-center text-base`,
-              ]}
-              onPress={handleGoogleSignIn}
-            >
-              <Image source={Google} style={tw``} />
-              <Text style={tw`mx-3 font-bold`}>Google</Text>
-            </TouchableOpacity>
-            <View
-              style={[
-                tw`py-4 px-6 shadow-md rounded-xl flex-1 w-full flex-row bg-white items-center text-base`,
-              ]}
-            >
-              <Image source={Fb} style={tw``} />
-              <Text style={tw`mx-3 font-bold`}>Facebook</Text>
-            </View>
+            <GoogleButton />
+            <FacebookButton />
           </View>
 
           <View style={tw`flex-row items-start w-full mt-6`}>
-            {/* <Text style={tw`mr-1`}>Don’t have an account?</Text> */}
             <Link
               href="/reset-password"
               style={tw`text-dark-gray underline my-4`}
