+++ types/env.d.tsx
@@ -3,4 +3,5 @@ declare module '@env' {
   export const WEB_CLIENT_ID: string;
   export const ANDROID_CLIENT_ID: string;
   export const IOS_CLIENT_ID: string;
+  export const FACEBOOK_APP_ID: string;
 }
